pyproject.toml
[project]
name = "motdle"
version = "0.1.0"
description = "Motdle - Un Wordle francais pour Discord"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.4.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
]

[project.scripts]
motdle = "motdle.__main__:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/motdle"]

[tool.pytest.ini_options]
testpaths = ["tests"]
